cleaning up the generated code to not break
commenting out a bunch

+++ bin/scanner.ml
@@ -1,4 +1,5 @@
 open Lang_agent    
+open Openai_client
 let window_size = ref 1024
     
 (* A function that returns the last line of a file that matches a given pattern *)
@@ -133,6 +134,7 @@ let traverse_and_print: 'client_t1 -> 'client_t2 -> string -> string -> string -
     let entries = Sys.readdir dir in
     Array.iter (fun entry ->
         let full_path = Filename.concat dir entry in
+        let full_out_path = full_path ^ ".out" in
         if Sys.is_directory full_path then
           (
             (print_endline ("DEBUG1:" ^  full_path));
@@ -150,49 +152,88 @@ let traverse_and_print: 'client_t1 -> 'client_t2 -> string -> string -> string -
             | S_DIR -> (print_endline ("DEBUG2 DIR" ^  full_path));
             | S_REG ->
                if Sys.file_exists  full_path then
-                 (* (print_endline ("DEBUG2 " ^  full_path)); *)
-                 let ic = open_in full_path in
 
-                 let chunks = split_file ic ! window_size in
-                 (* let ll =  last_line_matching ic in
-                    let chunks =  chunks_in ic in
-                  *)
-                 let do_one  (data)=
-                   let prompt = prompt1 ^ data in
-                   let res = (client1#lang_prompt param_record prompt ) in
-                   print_endline ("DEBUGINLINE@" ^ data ^ res);
-                   data ^ res
-                 in
-                 let ln = List.map do_one ! chunks  in
-                 let lr = List.rev ln in
-                 match lr  with
-                 | [] -> print_endline ("DEBUG@ERROR")
-                 | 
+                 if Sys.file_exists  full_out_path then
+                   print_endline ("SKIP" ^ full_out_path)
+                 else
+                   
+                   (* (print_endline ("DEBUG2 " ^  full_path)); *)
+                   let ic = open_in full_path in
+                   print_endline ("OPEN" ^ full_path);
+                   
+                   let chunks = split_file ic ! window_size in
+                   (* let ll =  last_line_matching ic in
+                      let chunks =  chunks_in ic in
+                    *)
+                   close_in ic;
+                   let do_one  (data)=
+                     let prompt = prompt1 ^ data in
+                     print_endline ("send" ^ prompt);
+                     let res = (client1#lang_prompt param_record prompt ) in
+                     print_endline ("get" ^ res);
 
-                   le::_ ->print_endline ("DEBUG20 " ^  String.concat "@" (parse_file_points le));
-                           ()
-                  
-        ) entries
-    in
-    aux path
+                     let oc = open_out full_out_path in
+                     
+                     print_endline ("DEBUGINLINE@" ^ data ^ res);
+                     Printf.fprintf oc "%s\n%s\n" data res;
+                     close_out oc;
+                     data ^ res
+                
+                   in
+                   let ln = List.map do_one ! chunks  in
+                   let lr = List.rev ln in
+                   match lr  with
+                   | [] -> print_endline ("DEBUG@ERROR")
+                   | 
+                     
+                     le::_ ->print_endline ("DEBUG20 " ^  String.concat "@" (parse_file_points le));
+                             ()
+                
+      ) entries
+  in
+  aux path
 
 let input_files = ref []
 
 let anon_fun filename =
   input_files := filename::!input_files
 
+module type M = sig
+  val test : unit -> unit
+end
+
+module M1 : M = struct
+  let init () = open_ai_lang_model.lang_init 
+end
+
+module M2 : M = struct
+  let init () = ollama_lang_model.lang_init 
+end
+
+
 let () =
   let start = ref "" in
   let prompt = ref "" in
   let model = ref "mistral" in
-  let binding = ref "ollama" in
+
+  let binding = ref (module M2 : M) in
+   
   let url = ref "" in
   let help_str = "test" in
   let opts = [
       "-s", Arg.Set_string start, "startdir";
       "-p", Arg.Set_string prompt, "prompt";
       "-m", Arg.Set_string model, "model";
-      "-b", Arg.Set_string binding, "binding";
+      "-bopenai", 
+      Arg.Unit (fun () ->
+          binding := (module M1 : M)), 
+      " Use Openai ";
+          
+      "-bollama", 
+      Arg.Unit (fun () ->
+          binding := (module M2 : M)),
+      " Use Ollama (default)"; 
+
       "-u", Arg.Set_string url, "url";
       "-w", Arg.Set_int window_size, "window_size";
   ] |> Arg.align in
@@ -200,24 +241,13 @@ let () =
   Arg.parse opts anon_fun help_str;
   Printf.printf "DEBUG3 Starting path %s\n" !start;
   (print_endline ("DEBUG4 MODEL :" ^ ! model) );
-  (print_endline ("DEBUG5 BINDING :" ^ ! binding) );
-  
-  if !binding == "openai" then
-    (print_endline ("DEBUG OPENAI :") );
-    let open_ai_client = new Openai_client.open_ai_lang_model in 
-    (* let client = Openai_client.create_init in *)
-    
-    (*Lang_model.client_t aka record of parameters
-     *)
-    let client_param_record = (open_ai_client#lang_init())  in
-
-    let client_param_record2 = open_ai_client#lang_open client_param_record ! url in
-    let client_param_record3 = open_ai_client#lang_set_model client_param_record2 ! model in
-    traverse_and_print open_ai_client client_param_record3 !start !model !prompt; 
-    (*         map_lookup_file_snippets fp snippet_size       *)
-    if !binding == "ollama" then
-      let ollama_client = new Openai_client.open_ai_lang_model in 
-      (* let client_old = Ollama.create_client ! model in *)
-      let client_param_record = (open_ai_client#lang_init())  in
-      traverse_and_print ollama_client client_param_record !start !model !prompt;         
-      (print_endline ("DEBUG1 :END") )
+ 
+  let open_ai_client = !binding in 
+  (* new Openai_client.open_ai_lang_model    ()   *)
+  (* Lang_model.client_t aka record of parameters *)
+   let client_param_record = (open_ai_client#lang_init())  in
+      
+      let client_param_record2 = open_ai_client#lang_open client_param_record ! url in
+      let client_param_record3 = open_ai_client#lang_set_model client_param_record2 ! model in
+      traverse_and_print open_ai_client client_param_record3 !start !model !prompt
+                         (*         map_lookup_file_snippets fp snippet_size       *)
